Workaround for bundler?/rgeo? option passing issue
Previously, passing in `--with-opt-dir` and `--with-geos-config` worked
fine. After upgrading my app to Ruby 2.6 and latest bundler, options
passed in through bundler don't seem to make it down to the C extension.

It's worth noting also, a couple years ago, rgeo changed what the
options were called, which resulted in heroku-geo-buildpack silently
breaking (PR #34).

This change puts `geos-config` and the misc build dependencies in
$PATH and $CPATH/$LD_LIBRARY_PATH/$LIBRARY_PATH as appropriate, so we
never have to deal with issues around passing arguments to bundler
again. Additionally, if you start a heroku bash prompt, you can fuss
with geos and building rgeo without learning the build options, which is
nice.